Move ThreatUpdateS3Store to s3_adapters.py
There is potential for re-use of the same class between indexer, matcher and fetcher. Having a single place for everything to do with signal storage makes refactors easier.
