+++ hasher-matcher-actioner/hmalib/common/s3_adapters.py
@@ -1,18 +1,35 @@
 # Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved
 
-import os
+import collections
+import csv
+import functools
+import warnings
+import io
+import json
 import boto3
 import csv
 import codecs
-
-import typing as t
+from pathlib import Path
+from datetime import datetime
 
 from dataclasses import dataclass, field
+import typing as t
+
+from botocore.errorfactory import ClientError
+from mypy_boto3_s3.service_resource import Bucket
+from threatexchange import threat_updates as tu
+from threatexchange.cli.dataset.simple_serialization import HMASerialization
+from threatexchange.descriptor import SimpleDescriptorRollup, ThreatDescriptor
+from threatexchange.signal_type.signal_base import SignalType
+
+
+from hmalib.common.signal_models import PDQSignalMetadata, PendingOpinionChange
 from hmalib import metrics
 from hmalib.common.logging import get_logger
 
 logger = get_logger(__name__)
 
+dynamodb = boto3.resource("dynamodb")
 s3_client = boto3.client("s3")
 HashRowT = t.Tuple[str, t.Dict[str, t.Any]]
 
@@ -153,3 +170,301 @@ class ThreatExchangeS3PDQAdapter(ThreatExchangeS3Adapter):
     @property
     def file_type_str_name(self):
         return "PDQ"
+
+
+class ThreatUpdateS3Store(tu.ThreatUpdatesStore):
+    """
+    ThreatUpdatesStore, but stores files in S3 instead of local filesystem.
+    """
+
+    def __init__(
+        self,
+        privacy_group: int,
+        app_id: int,
+        s3_bucket: Bucket,  # Not typable?
+        s3_te_data_folder: str,
+        data_store_table: str,
+        supported_signal_types: t.List[SignalType],
+    ) -> None:
+        super().__init__(privacy_group)
+        self.app_id = app_id
+        self._cached_state: t.Optional[t.Dict] = None
+        self.s3_bucket = s3_bucket
+        self.s3_te_data_folder = s3_te_data_folder
+        self.data_store_table = data_store_table
+        self.supported_signal_types = supported_signal_types
+
+    @functools.lru_cache(maxsize=None)
+    def _get_supported_indicator_types(self):
+        """
+        For supported self.signal_types, get their corresponding indicator_types.
+        """
+        indicator_types = [
+            "HASH_VIDEO_MD5"
+        ]  # hardcoding this because md5 currently points to HASH_MD5
+
+        for signal_type in self.supported_signal_types:
+            indicator_type = getattr(signal_type, "INDICATOR_TYPE", None)
+            if indicator_type:
+                indicator_types.append(indicator_type)
+            else:
+                warnings.warn(
+                    f"SignalType: {signal_type} does not provide an indicator type."
+                )
+
+        return indicator_types
+
+    @property
+    def checkpoint_s3_key(self) -> str:
+        return f"{self.s3_te_data_folder}{self.privacy_group}.checkpoint"
+
+    def get_s3_object_key(self, indicator_type) -> str:
+        """
+        Creates a 'file_name' structure that is dependent on downstream (see s3_adapters.py)
+        """
+        extension = f"{indicator_type.lower()}.te"
+        return f"{self.s3_te_data_folder}{self.privacy_group}.{extension}"
+
+    @property
+    def next_delta(self) -> tu.ThreatUpdatesDelta:
+        """
+        IF YOU CHANGE SUPPORTED_SIGNALS, OLD CHECKPOINTS NEED TO BE INVALIDATED
+        TO GET THE NON-PDQ DATA!
+        """
+        delta = super().next_delta
+        delta.types = self._get_supported_indicator_types()
+        return delta
+
+    def reset(self):
+        super().reset()
+        self._cached_state = None
+
+    def _load_checkpoint(self) -> tu.ThreatUpdateCheckpoint:
+        """Load the state of the threat_updates checkpoints from state directory"""
+        txt_content = read_s3_text(self.s3_bucket, self.checkpoint_s3_key)
+
+        if txt_content is None:
+            logger.warning("No s3 checkpoint for %d. First run?", self.privacy_group)
+            return tu.ThreatUpdateCheckpoint()
+        checkpoint_json = json.load(txt_content)
+
+        ret = tu.ThreatUpdateCheckpoint(
+            checkpoint_json["last_fetch_time"],
+            checkpoint_json["fetch_checkpoint"],
+        )
+        logger.info(
+            "Loaded checkpoint for privacy group %d. last_fetch_time=%d fetch_checkpoint=%d",
+            self.privacy_group,
+            ret.last_fetch_time,
+            ret.fetch_checkpoint,
+        )
+
+        return ret
+
+    def _store_checkpoint(self, checkpoint: tu.ThreatUpdateCheckpoint) -> None:
+        txt_content = io.StringIO()
+        json.dump(
+            {
+                "last_fetch_time": checkpoint.last_fetch_time,
+                "fetch_checkpoint": checkpoint.fetch_checkpoint,
+            },
+            txt_content,
+            indent=2,
+        )
+        write_s3_text(txt_content, self.s3_bucket, self.checkpoint_s3_key)
+        logger.info(
+            "Stored checkpoint for privacy group %d. last_fetch_time=%d fetch_checkpoint=%d",
+            self.privacy_group,
+            checkpoint.last_fetch_time,
+            checkpoint.fetch_checkpoint,
+        )
+
+    def load_state(self, allow_cached=True):
+        if not allow_cached or self._cached_state is None:
+            txt_content = read_s3_text(self.s3_bucket, self.data_s3_key)
+            items = []
+            if txt_content is None:
+                logger.warning("No TE state for %d. First run?", self.privacy_group)
+            else:
+                # Violate your warranty with module state!
+                csv.field_size_limit(65535)  # dodge field size problems
+                for row in csv.reader(txt_content):
+                    items.append(
+                        HMASerialization(
+                            row[0],
+                            "HASH_PDQ",
+                            row[1],
+                            SimpleDescriptorRollup.from_row(row[2:]),
+                        )
+                    )
+                logger.info("%d rows loaded for %d", len(items), self.privacy_group)
+            # Do all in one assignment just in case of threads
+            self._cached_state = {item.key: item for item in items}
+        return self._cached_state
+
+    def _store_state(self, contents: t.Iterable["HMASerialization"]):
+        row_by_type: t.DefaultDict = collections.defaultdict(list)
+        for item in contents:
+            row_by_type[item.indicator_type].append(item)
+        # Discard all updates except PDQ
+
+        for indicator_type in row_by_type:
+            # Write one file per indicator type.
+            items = row_by_type.get(indicator_type, [])
+
+            with io.StringIO(newline="") as txt_content:
+                writer = csv.writer(txt_content)
+                writer.writerows(item.as_csv_row() for item in items)
+
+                write_s3_text(
+                    txt_content, self.s3_bucket, self.get_s3_object_key(indicator_type)
+                )
+
+                logger.info(
+                    "IndicatorType:%s, %d rows stored in PrivacyGroup %d",
+                    indicator_type,
+                    len(items),
+                    self.privacy_group,
+                )
+
+    def _apply_updates_impl(
+        self,
+        delta: tu.ThreatUpdatesDelta,
+        post_apply_fn=lambda x: None,
+    ) -> None:
+        state: t.Dict = {}
+        updated: t.Dict = {}
+        if delta.start > 0:
+            state = self.load_state()
+        for update in delta:
+            item = HMASerialization.from_threat_updates_json(
+                self.app_id, update.raw_json
+            )
+            if update.should_delete:
+                state.pop(item.key, None)
+            else:
+                state[item.key] = item
+                updated[item.key] = item
+
+        self._store_state(state.values())
+        self._cached_state = state
+
+        post_apply_fn(updated)
+
+    def get_new_pending_opinion_change(
+        self, metadata: PDQSignalMetadata, new_tags: t.List[str]
+    ):
+        # Figure out if we have a new opinion about this indicator and clear out a pending change if so
+
+        # python-threatexchange.descriptor.ThreatDescriptor.from_te_json guarentees there is either
+        # 0 or 1 opinion tags on a descriptor
+        opinion_tags = ThreatDescriptor.SPECIAL_TAGS
+        old_opinion = [tag for tag in metadata.tags if tag in opinion_tags]
+        new_opinion = [tag for tag in new_tags if tag in opinion_tags]
+
+        # If our opinion changed or if our pending change has already happend,
+        # set the pending opinion change to None, otherwise keep it unchanged
+        if old_opinion != new_opinion:
+            return PendingOpinionChange.NONE
+        elif (
+            (
+                new_opinion == [ThreatDescriptor.TRUE_POSITIVE]
+                and metadata.pending_opinion_change
+                == PendingOpinionChange.MARK_TRUE_POSITIVE
+            )
+            or (
+                new_opinion == [ThreatDescriptor.FALSE_POSITIVE]
+                and metadata.pending_opinion_change
+                == PendingOpinionChange.MARK_FALSE_POSITIVE
+            )
+            or (
+                new_opinion == []
+                and metadata.pending_opinion_change
+                == PendingOpinionChange.REMOVE_OPINION
+            )
+        ):
+            return PendingOpinionChange.NONE
+        else:
+            return metadata.pending_opinion_change
+
+    def post_apply(self, updated: t.Dict = {}):
+        """
+        After the fetcher applies an update, check for matches
+        to any of the signals in data_store_table and if found update
+        their tags.
+
+        TODO: Additionally, if writebacks are enabled for this privacy group write back
+        INGESTED to ThreatExchange
+        """
+        table = dynamodb.Table(self.data_store_table)
+
+        for update in updated.values():
+            if update.indicator_type == "HASH_PDQ":
+                # To remain backwards compatible, only handle PDQ here. SME:
+                # @schatten, @barrettolson
+
+                # TODO: Once Signal storage is generified, remove this
+                # conditional and handle multiple indicator_types.
+
+                row: t.List[str] = update.as_csv_row()
+                # example row format: ('<raw_indicator>', '<indicator-id>', '<descriptor-id>', '<time added>', '<space-separated-tags>')
+                # e.g (10736405276340','096a6f9...064f', '1234567890', '2020-07-31T18:47:45+0000', 'true_positive hma_test')
+                new_tags = row[4].split(" ") if row[4] else []
+
+                metadata = PDQSignalMetadata.get_from_signal_and_ds_id(
+                    table,
+                    int(row[1]),
+                    S3ThreatDataConfig.SOURCE_STR,
+                    str(self.privacy_group),
+                )
+
+                if metadata:
+                    new_pending_opinion_change = self.get_new_pending_opinion_change(
+                        metadata, new_tags
+                    )
+                else:
+                    # If this is a new indicator without metadata there is nothing for us to update
+                    return
+
+                metadata = PDQSignalMetadata(
+                    signal_id=row[1],
+                    ds_id=str(self.privacy_group),
+                    updated_at=datetime.now(),
+                    signal_source=S3ThreatDataConfig.SOURCE_STR,
+                    signal_hash=row[
+                        0
+                    ],  # note: not used by update_tags_in_table_if_exists
+                    tags=new_tags,
+                    pending_opinion_change=new_pending_opinion_change,
+                )
+                # TODO: Combine 2 update functions into single function
+                if metadata.update_tags_in_table_if_exists(table):
+                    logger.info(
+                        "Updated Signal Tags in DB for indicator id: %s source: %s for privacy group: %d",
+                        row[1],
+                        S3ThreatDataConfig.SOURCE_STR,
+                        self.privacy_group,
+                    )
+                if metadata.update_pending_opinion_change_in_table_if_exists(table):
+                    logger.info(
+                        "Updated Pending Opinion in DB for indicator id: %s source: %s for privacy group: %d",
+                        row[1],
+                        S3ThreatDataConfig.SOURCE_STR,
+                        self.privacy_group,
+                    )
+
+
+def read_s3_text(bucket, key: str) -> t.Optional[io.StringIO]:
+    byte_content = io.BytesIO()
+    try:
+        bucket.download_fileobj(key, byte_content)
+    except ClientError as ce:
+        if ce.response["Error"]["Code"] != "404":
+            raise
+        return None
+    return io.StringIO(byte_content.getvalue().decode())
+
+
+def write_s3_text(txt_content: io.StringIO, bucket, key: str) -> None:
+    byte_content = io.BytesIO(txt_content.getvalue().encode())
+    bucket.upload_fileobj(byte_content, key)
